预览，编译抽离
